+++ cognite/neat/rules/models/asset/_converter.py
@@ -1,4 +1,0 @@
-from cognite.neat.rules.transformers._converters import _InformationRulesConverter
-
-
-class _AssetRulesConverter(_InformationRulesConverter): ...