+++ cognite/neat/rules/models/asset/_rules.py
@@ -7,7 +7,6 @@ from rdflib import Namespace
 
 from cognite.neat.constants import get_default_prefixes
 from cognite.neat.rules.models._base import BaseRules, RoleTypes, SheetList
-from cognite.neat.rules.models.domain import DomainRules
 from cognite.neat.rules.models.entities import (
     CdfResourceEntityList,
     ClassEntity,
@@ -22,7 +21,7 @@ from cognite.neat.rules.models.information import (
 )
 
 if TYPE_CHECKING:
-    from cognite.neat.rules.models.dms._rules import DMSRules
+    pass
 
 
 if sys.version_info >= (3, 11):
@@ -141,16 +140,3 @@ class AssetRules(BaseRules):
             prefix = self.reference.metadata.prefix
             cleaned[reference] = _AssetRulesSerializer(by_alias, prefix).clean(ref_dump, True)
         return cleaned
-
-    def as_domain_rules(self) -> DomainRules:
-        from ._converter import _AssetRulesConverter
-
-        return _AssetRulesConverter(self.as_information_rules()).as_domain_rules()
-
-    def as_dms_rules(self) -> "DMSRules":
-        from ._converter import _AssetRulesConverter
-
-        return _AssetRulesConverter(self.as_information_rules()).as_dms_rules()
-
-    def as_information_rules(self) -> InformationRules:
-        return InformationRules.model_validate(self.model_dump())