+++ cognite/neat/rules/models/information/_rules.py
@@ -34,7 +34,6 @@ from cognite.neat.rules.models._types import (
     VersionType,
 )
 from cognite.neat.rules.models.data_types import DataType
-from cognite.neat.rules.models.domain import DomainRules
 from cognite.neat.rules.models.entities import (
     ClassEntity,
     ClassEntityList,
@@ -47,7 +46,7 @@ from cognite.neat.rules.models.entities import (
 )
 
 if TYPE_CHECKING:
-    from cognite.neat.rules.models import AssetRules, DMSRules
+    from cognite.neat.rules.models import DMSRules
 
 
 if sys.version_info >= (3, 11):
@@ -339,16 +338,6 @@ class InformationRules(BaseRules):
             prefix = self.reference.metadata.prefix
             cleaned[reference] = _InformationRulesSerializer(by_alias, prefix).clean(ref_dump, True)
         return cleaned
-
-    def as_domain_rules(self) -> DomainRules:
-        from cognite.neat.rules.transformers._converters import _InformationRulesConverter
-
-        return _InformationRulesConverter(self).as_domain_rules()
-
-    def as_asset_architect_rules(self) -> "AssetRules":
-        from cognite.neat.rules.transformers._converters import _InformationRulesConverter
-
-        return _InformationRulesConverter(self).as_asset_architect_rules()
 
     def as_dms_rules(self) -> "DMSRules":
         from cognite.neat.rules.transformers._converters import _InformationRulesConverter